Add the GUI for the Rock Paper Scissors game.
I added css to style the user interface of the game and make it more beautiful. I changed the selections for rock, paper, and scissors to
images nested in buttons rather than obtaining the input from a prompt. I attached the string value to each of the buttons and added an
event listener to listen for when the user clicks the button, and use the value for that button to run the game, and also display the
result of the game in a div.