Add the score between the players
I created two variables that keeps track of the score of the player and computer and displays the result in a div. I also wrote a conditional
statement to announce the winner of the round when a player reaches 5 points. Then I added more styles to the webpage and made the user interface
more responsive.